Introduce StateChanged 'event' with basic implementation
Also extract HeadState to own module
