+++ hydra-node/src/Hydra/HeadLogic.hs
@@ -42,8 +42,18 @@ import Hydra.Crypto (
   sign,
   verifyMultiSignature,
  )
+import Hydra.HeadLogic.HeadState (
+  ClosedState (..),
+  CoordinatedHeadState (..),
+  HeadState (..),
+  IdleState (..),
+  InitialState (..),
+  OpenState (..),
+  SeenSnapshot (..),
+  StateChanged (..),
+  Committed, seenSnapshotNumber, setChainState
+ )
 import Hydra.Ledger (
-  ChainSlot,
   IsTx,
   Ledger (..),
   UTxOType,
@@ -114,207 +124,6 @@ instance
   where
   arbitrary = genericArbitrary
 
--- | The main state of the Hydra protocol state machine. It holds both, the
--- overall protocol state, but also the off-chain 'CoordinatedHeadState'.
---
--- Each of the sub-types (InitialState, OpenState, etc.) contain black-box
--- 'chainState' corresponding to 'OnChainEvent' that has been observed leading
--- to the state.
---
--- Note that rollbacks are currently not fully handled in the head logic and
--- only this internal chain state gets replaced with the "rolled back to"
--- version.
-data HeadState tx
-  = Idle (IdleState tx)
-  | Initial (InitialState tx)
-  | Open (OpenState tx)
-  | Closed (ClosedState tx)
-  deriving stock (Generic)
-
-instance (IsTx tx, Arbitrary (ChainStateType tx)) => Arbitrary (HeadState tx) where
-  arbitrary = genericArbitrary
-
-deriving instance (IsTx tx, Eq (ChainStateType tx)) => Eq (HeadState tx)
-deriving instance (IsTx tx, Show (ChainStateType tx)) => Show (HeadState tx)
-deriving instance (IsTx tx, ToJSON (ChainStateType tx)) => ToJSON (HeadState tx)
-deriving instance (IsTx tx, FromJSON (ChainStateType tx)) => FromJSON (HeadState tx)
-
--- | Get the chain state in any 'HeadState'.
-getChainState :: HeadState tx -> ChainStateType tx
-getChainState = \case
-  Idle IdleState{chainState} -> chainState
-  Initial InitialState{chainState} -> chainState
-  Open OpenState{chainState} -> chainState
-  Closed ClosedState{chainState} -> chainState
-
--- | Update the chain state in any 'HeadState'.
-setChainState :: ChainStateType tx -> HeadState tx -> HeadState tx
-setChainState chainState = \case
-  Idle st -> Idle st{chainState}
-  Initial st -> Initial st{chainState}
-  Open st -> Open st{chainState}
-  Closed st -> Closed st{chainState}
-
--- ** Idle
-
--- | An 'Idle' head only having a chain state with things seen on chain so far.
-newtype IdleState tx = IdleState {chainState :: ChainStateType tx}
-  deriving (Generic)
-
-deriving instance Eq (ChainStateType tx) => Eq (IdleState tx)
-deriving instance Show (ChainStateType tx) => Show (IdleState tx)
-deriving anyclass instance ToJSON (ChainStateType tx) => ToJSON (IdleState tx)
-deriving anyclass instance FromJSON (ChainStateType tx) => FromJSON (IdleState tx)
-
-instance (Arbitrary (ChainStateType tx)) => Arbitrary (IdleState tx) where
-  arbitrary = genericArbitrary
-
--- ** Initial
-
--- | An 'Initial' head which already has an identity and is collecting commits.
-data InitialState tx = InitialState
-  { parameters :: HeadParameters
-  , pendingCommits :: PendingCommits
-  , committed :: Committed tx
-  , chainState :: ChainStateType tx
-  , headId :: HeadId
-  }
-  deriving (Generic)
-
-deriving instance (IsTx tx, Eq (ChainStateType tx)) => Eq (InitialState tx)
-deriving instance (IsTx tx, Show (ChainStateType tx)) => Show (InitialState tx)
-deriving instance (IsTx tx, ToJSON (ChainStateType tx)) => ToJSON (InitialState tx)
-deriving instance (IsTx tx, FromJSON (ChainStateType tx)) => FromJSON (InitialState tx)
-
-instance (IsTx tx, Arbitrary (ChainStateType tx)) => Arbitrary (InitialState tx) where
-  arbitrary = do
-    InitialState
-      <$> arbitrary
-      <*> arbitrary
-      <*> arbitrary
-      <*> arbitrary
-      <*> arbitrary
-
-type PendingCommits = Set Party
-
-type Committed tx = Map Party (UTxOType tx)
-
--- ** Open
-
--- | An 'Open' head with a 'CoordinatedHeadState' tracking off-chain
--- transactions.
-data OpenState tx = OpenState
-  { parameters :: HeadParameters
-  , coordinatedHeadState :: CoordinatedHeadState tx
-  , chainState :: ChainStateType tx
-  , headId :: HeadId
-  , currentSlot :: ChainSlot
-  }
-  deriving (Generic)
-
-deriving instance (IsTx tx, Eq (ChainStateType tx)) => Eq (OpenState tx)
-deriving instance (IsTx tx, Show (ChainStateType tx)) => Show (OpenState tx)
-deriving instance (IsTx tx, ToJSON (ChainStateType tx)) => ToJSON (OpenState tx)
-deriving instance (IsTx tx, FromJSON (ChainStateType tx)) => FromJSON (OpenState tx)
-
-instance (IsTx tx, Arbitrary (ChainStateType tx)) => Arbitrary (OpenState tx) where
-  arbitrary =
-    OpenState
-      <$> arbitrary
-      <*> arbitrary
-      <*> arbitrary
-      <*> arbitrary
-      <*> arbitrary
-
--- | Off-chain state of the Coordinated Head protocol.
-data CoordinatedHeadState tx = CoordinatedHeadState
-  { seenUTxO :: UTxOType tx
-  -- ^ The latest UTxO resulting from applying 'seenTxs' to
-  -- 'confirmedSnapshot'. Spec: L̂
-  , seenTxs :: [tx]
-  -- ^ List of seen transactions pending inclusion in a snapshot. Spec: T̂
-  , confirmedSnapshot :: ConfirmedSnapshot tx
-  -- ^ The latest confirmed snapshot. Spec: U̅, s̅ and σ̅
-  , seenSnapshot :: SeenSnapshot tx
-  -- ^ Last seen snapshot and signatures accumulator. Spec: Û, ŝ and Σ̂
-  }
-  deriving stock (Generic)
-
-deriving instance IsTx tx => Eq (CoordinatedHeadState tx)
-deriving instance IsTx tx => Show (CoordinatedHeadState tx)
-deriving instance IsTx tx => ToJSON (CoordinatedHeadState tx)
-deriving instance IsTx tx => FromJSON (CoordinatedHeadState tx)
-
-instance IsTx tx => Arbitrary (CoordinatedHeadState tx) where
-  arbitrary = genericArbitrary
-
--- | Data structure to help in tracking whether we have seen or requested a
--- ReqSn already and if seen, the signatures we collected already.
-data SeenSnapshot tx
-  = -- | Never saw a ReqSn.
-    NoSeenSnapshot
-  | -- | No snapshot in flight with last seen snapshot number as given.
-    LastSeenSnapshot {lastSeen :: SnapshotNumber}
-  | -- | ReqSn was sent out and it should be considered already in flight.
-    RequestedSnapshot
-      { lastSeen :: SnapshotNumber
-      , requested :: SnapshotNumber
-      }
-  | -- | ReqSn for given snapshot was received.
-    SeenSnapshot
-      { snapshot :: Snapshot tx
-      , signatories :: Map Party (Signature (Snapshot tx))
-      -- ^ Collected signatures and so far.
-      }
-  deriving stock (Generic)
-
-instance IsTx tx => Arbitrary (SeenSnapshot tx) where
-  arbitrary = genericArbitrary
-
-deriving instance IsTx tx => Eq (SeenSnapshot tx)
-deriving instance IsTx tx => Show (SeenSnapshot tx)
-deriving instance IsTx tx => ToJSON (SeenSnapshot tx)
-deriving instance IsTx tx => FromJSON (SeenSnapshot tx)
-
--- | Get the last seen snapshot number given a 'SeenSnapshot'.
-seenSnapshotNumber :: SeenSnapshot tx -> SnapshotNumber
-seenSnapshotNumber = \case
-  NoSeenSnapshot -> 0
-  LastSeenSnapshot{lastSeen} -> lastSeen
-  RequestedSnapshot{lastSeen} -> lastSeen
-  SeenSnapshot{snapshot = Snapshot{number}} -> number
-
--- ** Closed
-
--- | An 'Closed' head with an current candidate 'ConfirmedSnapshot', which may
--- be contested before the 'contestationDeadline'.
-data ClosedState tx = ClosedState
-  { parameters :: HeadParameters
-  , confirmedSnapshot :: ConfirmedSnapshot tx
-  , contestationDeadline :: UTCTime
-  , readyToFanoutSent :: Bool
-  -- ^ Tracks whether we have informed clients already about being
-  -- 'ReadyToFanout'.
-  , chainState :: ChainStateType tx
-  , headId :: HeadId
-  }
-  deriving (Generic)
-
-deriving instance (IsTx tx, Eq (ChainStateType tx)) => Eq (ClosedState tx)
-deriving instance (IsTx tx, Show (ChainStateType tx)) => Show (ClosedState tx)
-deriving instance (IsTx tx, ToJSON (ChainStateType tx)) => ToJSON (ClosedState tx)
-deriving instance (IsTx tx, FromJSON (ChainStateType tx)) => FromJSON (ClosedState tx)
-
-instance (IsTx tx, Arbitrary (ChainStateType tx)) => Arbitrary (ClosedState tx) where
-  arbitrary =
-    ClosedState
-      <$> arbitrary
-      <*> arbitrary
-      <*> arbitrary
-      <*> arbitrary
-      <*> arbitrary
-      <*> arbitrary
-
 -- ** Other types
 
 type TTL = Natural
@@ -357,7 +166,7 @@ instance Arbitrary RequirementFailure where
 
 data Outcome tx
   = OnlyEffects {effects :: [Effect tx]}
-  | NewState {headState :: HeadState tx, effects :: [Effect tx]}
+  | NewState {headState :: StateChanged tx, effects :: [Effect tx]}
   | Wait {reason :: WaitReason}
   | Error {error :: LogicError tx}
   deriving stock (Generic)
@@ -426,16 +235,19 @@ onIdleChainInitTx ::
   Outcome tx
 onIdleChainInitTx newChainState parties contestationPeriod headId =
   NewState
-    ( Initial
-        InitialState
-          { parameters = HeadParameters{contestationPeriod, parties}
-          , pendingCommits = Set.fromList parties
-          , committed = mempty
-          , chainState = newChainState
-          , headId
-          }
-    )
-    [ClientEffect $ HeadIsInitializing headId (fromList parties)]
+    { headState =
+        StateChanged
+          ( Initial
+              InitialState
+                { parameters = HeadParameters{contestationPeriod, parties}
+                , pendingCommits = Set.fromList parties
+                , committed = mempty
+                , chainState = newChainState
+                , headId
+                }
+          )
+    , effects = [ClientEffect $ HeadIsInitializing headId (fromList parties)]
+    }
 
 -- | Client request to commit a UTxO entry to the head. Provided the client
 -- hasn't committed yet, this leads to a commit transaction on-chain containing
@@ -477,7 +289,7 @@ onInitialChainCommitTx ::
   UTxOType tx ->
   Outcome tx
 onInitialChainCommitTx st newChainState pt utxo =
-  NewState newState $
+  NewState (StateChanged newState) $
     notifyClient
       : [postCollectCom | canCollectCom]
  where
@@ -532,7 +344,7 @@ onInitialChainAbortTx ::
   Outcome tx
 onInitialChainAbortTx newChainState committed headId =
   NewState
-    (Idle IdleState{chainState = newChainState})
+    (StateChanged $ Idle IdleState{chainState = newChainState})
     [ClientEffect $ HeadIsAborted{headId, utxo = fold committed}]
 
 -- | Observe a collectCom transaction. We initialize the 'OpenState' using the
@@ -548,7 +360,12 @@ onInitialChainCollectTx ::
   Outcome tx
 onInitialChainCollectTx st newChainState =
   NewState
-    ( Open
+    newState
+    [ClientEffect $ HeadIsOpen{headId, utxo = u0}]
+ where
+  newState =
+    StateChanged $
+      Open
         OpenState
           { parameters
           , coordinatedHeadState =
@@ -557,9 +374,7 @@ onInitialChainCollectTx st newChainState =
           , headId
           , currentSlot = chainStateSlot newChainState
           }
-    )
-    [ClientEffect $ HeadIsOpen{headId, utxo = u0}]
- where
+
   u0 = fold committed
 
   initialSnapshot = InitialSnapshot u0
@@ -609,15 +424,7 @@ onOpenNetworkReqTx env ledger st ttl tx =
       | otherwise -> Wait $ WaitOnNotApplicableTx err
     Right utxo' ->
       NewState
-        ( Open
-            st
-              { coordinatedHeadState =
-                  coordinatedHeadState
-                    { seenTxs = seenTxs <> [tx]
-                    , seenUTxO = utxo'
-                    }
-              }
-        )
+        (newState utxo')
         [ClientEffect $ TxValid headId tx]
         & emitSnapshot env
  where
@@ -626,6 +433,18 @@ onOpenNetworkReqTx env ledger st ttl tx =
   CoordinatedHeadState{seenTxs, seenUTxO} = coordinatedHeadState
 
   OpenState{coordinatedHeadState, headId, currentSlot} = st
+
+  newState utxo' =
+    StateChanged
+      ( Open
+          st
+            { coordinatedHeadState =
+                coordinatedHeadState
+                  { seenTxs = seenTxs <> [tx]
+                  , seenUTxO = utxo'
+                  }
+            }
+      )
 
 -- | Process a snapshot request ('ReqSn') from party.
 --
@@ -668,17 +487,20 @@ onOpenNetworkReqSn env ledger st otherParty sn requestedTxs =
         -- Spec: T̂ ← {tx | ∀tx ∈ T̂ , Û ◦ tx ≠ ⊥} and L̂ ← Û ◦ T̂
         let (seenTxs', seenUTxO') = pruneTransactions u
         NewState
-          ( Open
-              st
-                { coordinatedHeadState =
-                    coordinatedHeadState
-                      { seenSnapshot = SeenSnapshot nextSnapshot mempty
-                      , seenTxs = seenTxs'
-                      , seenUTxO = seenUTxO'
+          { headState =
+              StateChanged
+                ( Open
+                    st
+                      { coordinatedHeadState =
+                          coordinatedHeadState
+                            { seenSnapshot = SeenSnapshot nextSnapshot mempty
+                            , seenTxs = seenTxs'
+                            , seenUTxO = seenUTxO'
+                            }
                       }
-                }
-          )
-          [NetworkEffect $ AckSn party snapshotSignature sn]
+                )
+          , effects = [NetworkEffect $ AckSn party snapshotSignature sn]
+          }
  where
   requireReqSn continue =
     if
@@ -764,17 +586,20 @@ onOpenNetworkAckSn env openState otherParty snapshotSignature sn =
           let multisig = aggregateInOrder sigs' parties
           requireVerifiedMultisignature multisig snapshot $ do
             NewState
-              ( onlyUpdateCoordinatedHeadState $
-                  coordinatedHeadState
-                    { confirmedSnapshot =
-                        ConfirmedSnapshot
-                          { snapshot
-                          , signatures = multisig
+              { headState =
+                  StateChanged
+                    ( onlyUpdateCoordinatedHeadState $
+                        coordinatedHeadState
+                          { confirmedSnapshot =
+                              ConfirmedSnapshot
+                                { snapshot
+                                , signatures = multisig
+                                }
+                          , seenSnapshot = LastSeenSnapshot (number snapshot)
                           }
-                    , seenSnapshot = LastSeenSnapshot (number snapshot)
-                    }
-              )
-              [ClientEffect $ SnapshotConfirmed headId snapshot multisig]
+                    )
+              , effects = [ClientEffect $ SnapshotConfirmed headId snapshot multisig]
+              }
               & emitSnapshot env
  where
   seenSn = seenSnapshotNumber seenSnapshot
@@ -800,12 +625,15 @@ onOpenNetworkAckSn env openState otherParty snapshotSignature sn =
       then cont
       else
         NewState
-          ( onlyUpdateCoordinatedHeadState $
-              coordinatedHeadState
-                { seenSnapshot = SeenSnapshot snapshot sigs'
-                }
-          )
-          []
+          { headState =
+              StateChanged
+                ( onlyUpdateCoordinatedHeadState $
+                    coordinatedHeadState
+                      { seenSnapshot = SeenSnapshot snapshot sigs'
+                      }
+                )
+          , effects = []
+          }
 
   requireVerifiedMultisignature multisig msg cont =
     if verifyMultiSignature vkeys multisig msg
@@ -869,15 +697,16 @@ onOpenChainCloseTx openState newChainState closedSnapshotNumber contestationDead
     number (getSnapshot confirmedSnapshot) > closedSnapshotNumber
 
   closedState =
-    Closed
-      ClosedState
-        { parameters
-        , confirmedSnapshot
-        , contestationDeadline
-        , readyToFanoutSent = False
-        , chainState = newChainState
-        , headId
-        }
+    StateChanged $
+      Closed
+        ClosedState
+          { parameters
+          , confirmedSnapshot
+          , contestationDeadline
+          , readyToFanoutSent = False
+          , chainState = newChainState
+          , headId
+          }
 
   notifyClient =
     ClientEffect $
@@ -944,7 +773,7 @@ onClosedChainFanoutTx ::
   Outcome tx
 onClosedChainFanoutTx closedState newChainState =
   NewState
-    (Idle IdleState{chainState = newChainState})
+    (StateChanged $ Idle IdleState{chainState = newChainState})
     [ ClientEffect $ HeadIsFinalized{headId, utxo}
     ]
  where
@@ -995,10 +824,11 @@ update env ledger st ev = case (st, ev) of
     onOpenNetworkAckSn env openState otherParty snapshotSignature sn
   ( Open openState@OpenState{headId = ourHeadId}
     , OnChainEvent Observation{observedTx = OnCloseTx{headId, snapshotNumber = closedSnapshotNumber, contestationDeadline}, newChainState}
-    ) | ourHeadId == headId ->
-      onOpenChainCloseTx openState newChainState closedSnapshotNumber contestationDeadline
+    )
+      | ourHeadId == headId ->
+          onOpenChainCloseTx openState newChainState closedSnapshotNumber contestationDeadline
       | otherwise ->
-        Error NotOurHead{ourHeadId, otherHeadId = headId}
+          Error NotOurHead{ourHeadId, otherHeadId = headId}
   (Open OpenState{coordinatedHeadState = CoordinatedHeadState{confirmedSnapshot}, headId}, ClientEvent GetUTxO) ->
     -- TODO: Is it really intuitive that we respond from the confirmed ledger if
     -- transactions are validated against the seen ledger?
@@ -1009,7 +839,7 @@ update env ledger st ev = case (st, ev) of
   (Closed cst@ClosedState{contestationDeadline, readyToFanoutSent, headId}, OnChainEvent Tick{chainTime})
     | chainTime > contestationDeadline && not readyToFanoutSent ->
         NewState
-          (Closed cst{readyToFanoutSent = True})
+          (StateChanged $ Closed cst{readyToFanoutSent = True})
           [ClientEffect $ ReadyToFanout headId]
   (Closed closedState, ClientEvent Fanout) ->
     onClosedClientFanout closedState
@@ -1017,9 +847,9 @@ update env ledger st ev = case (st, ev) of
     onClosedChainFanoutTx closedState newChainState
   -- General
   (currentState, OnChainEvent Rollback{rolledBackChainState}) ->
-    NewState (setChainState rolledBackChainState currentState) []
+    NewState (StateChanged $ setChainState rolledBackChainState currentState) []
   (Open ost@OpenState{}, OnChainEvent Tick{chainSlot}) ->
-    NewState (Open ost{currentSlot = chainSlot}) []
+    NewState (StateChanged $ Open ost{currentSlot = chainSlot}) []
   (_, OnChainEvent Tick{}) ->
     OnlyEffects []
   (_, NetworkEvent _ (Connected nodeId)) ->
@@ -1084,26 +914,27 @@ newSn Environment{party} parameters CoordinatedHeadState{confirmedSnapshot, seen
 emitSnapshot :: Environment -> Outcome tx -> Outcome tx
 emitSnapshot env@Environment{party} outcome =
   case outcome of
-    NewState (Open OpenState{parameters, coordinatedHeadState, chainState, headId, currentSlot}) effects ->
+    NewState (StateChanged (Open OpenState{parameters, coordinatedHeadState, chainState, headId, currentSlot})) effects ->
       case newSn env parameters coordinatedHeadState of
         ShouldSnapshot sn txs -> do
           let CoordinatedHeadState{seenSnapshot} = coordinatedHeadState
           NewState
-            ( Open
-                OpenState
-                  { parameters
-                  , coordinatedHeadState =
-                      coordinatedHeadState
-                        { seenSnapshot =
-                            RequestedSnapshot
-                              { lastSeen = seenSnapshotNumber seenSnapshot
-                              , requested = sn
-                              }
-                        }
-                  , chainState
-                  , headId
-                  , currentSlot
-                  }
+            ( StateChanged $
+                Open
+                  OpenState
+                    { parameters
+                    , coordinatedHeadState =
+                        coordinatedHeadState
+                          { seenSnapshot =
+                              RequestedSnapshot
+                                { lastSeen = seenSnapshotNumber seenSnapshot
+                                , requested = sn
+                                }
+                          }
+                    , chainState
+                    , headId
+                    , currentSlot
+                    }
             )
             $ NetworkEffect (ReqSn party sn txs) : effects
         _ -> outcome
